+++ src/components/Contact.css
@@ -1,0 +1,40 @@
+.Contact {
+    max-width: 400px;
+    margin: 0 auto;
+    text-align: center;
+
+}
+
+.Contact label {
+    text-align: center;
+    display: block;
+}
+
+.Contact h3 {
+    font-size: 20px;
+    color: black;
+    margin-bottom: 30px;
+}
+
+.Contact input,
+.Contact textarea {
+    width: 100%;
+    padding: 6px 10px;
+    margin: 10px 0;
+    border: 1px solid #ddd;
+    box-sizing: border-box;
+    display: block;
+    border-radius: 8px;
+    font-family: 'Roboto',
+        sans-serif;
+}
+
+.Contact button {
+    background: #f1356d;
+    color: #fff;
+    border: 0;
+    padding: 8px;
+    border-radius: 8px;
+    cursor: pointer;
+
+}
