+++ src/components/Footer.css
@@ -32,7 +32,9 @@
     width: 100%;
 
 }
-
+p {
+    color:white;
+}
 
 
 ul{
